Group tests by namespaces
