+++ phpcs.xml
@@ -7,5 +7,6 @@
     <exclude-pattern>resources</exclude-pattern>
     <exclude-pattern>database/</exclude-pattern>
     <exclude-pattern>storage/</exclude-pattern>
+    <exclude-pattern>tests/</exclude-pattern>
     <exclude-pattern>node_modules/</exclude-pattern>
 </ruleset>
