+++ tests/Feature/Jetstream/EmailVerificationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Jetstream;
 
 use App\Models\User;
 use App\Providers\RouteServiceProvider;
