+++ tests/Feature/Jetstream/ExampleTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Tests\Feature\Jetstream;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
